call context as guardrail inner interface
call context as guardrail inner interface
